rename lib-redux to slice-factory
also when i brought my changes to my branch it lost the refactor renames so now it looks like i deleted files and created files.

The only changes i made were to rename the package, to the readmes, and i created the mapErrorToSerializableObject function
