+++ packages/slice-factory/.eslintrc
