+++ packages/slice-factory/src/index.ts
@@ -1,11 +1,11 @@
 import StateStatusEnum from './constants/StateStatusEnum';
-import createEntitySlice from './higher-order-slices/create-entity-slice';
-import type { IEntitySlice, IEntitySliceReducers, IEntitySliceSelectors } from './higher-order-slices/create-entity-slice';
-import createModelSlice from './higher-order-slices/create-model-slice';
-import type { IModelSlice, IModelSliceReducers, IModelSliceSelectors } from './higher-order-slices/create-model-slice';
 import { IEntityState } from './models/entity-state';
 import { IMetaState } from './models/meta-state';
 import { IModelState } from './models/model-state';
+import createEntitySlice from './slice-factories/create-entity-slice';
+import type { IEntitySlice, IEntitySliceReducers, IEntitySliceSelectors } from './slice-factories/create-entity-slice';
+import createModelSlice from './slice-factories/create-model-slice';
+import type { IModelSlice, IModelSliceReducers, IModelSliceSelectors } from './slice-factories/create-model-slice';
 import { ISlice, ISliceSelectors, IMetaSliceSelectors } from './types';
 
 export {
