+++ packages/slice-factory/src/models/entity-state/EntityState.ts
