+++ packages/slice-factory/src/models/model-state/ModelState.ts
