+++ packages/slice-factory/src/slice-factories/create-model-slice/index.ts
