+++ packages/slice-factory/src/types/index.ts
